Time: 4 ms (82.57%), Space: 42 MB (68.6%) - LeetHub

+++ 0633-sum-of-square-numbers/0633-sum-of-square-numbers.java
@@ -4,7 +4,7 @@ class Solution {
         //     return true;
         // }
         long num = (long)Math.ceil(Math.sqrt(c));
-        int low = 0;
+        long low = 0;
         long high = num;
         while(low <= high){
             long sum = (low*low) + (high*high);
